% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house-loading.R
\name{load_house_experiments}
\alias{load_house_experiments}
\title{Loads supermarket experiment data from folder}
\usage{
load_house_experiments(folder)
}
\arguments{
\item{folder}{folder with one or more supermarket experiments}

\item{language}{language of the items in the experiment log.
Only important if you are not logging item codes
See language options in [item_categories]. Default is "CZ".}
}
\value{
list with loaded experiments
}
\description{
Loads supermarket experiment data from folder
}
